nft展示、mint、链接钱包等

+++ scripts/deploy.js
@@ -2,6 +2,7 @@
 // yours, or create new ones.
 
 const path = require("path");
+const fs = require("fs");
 
 async function main() {
   // This is just a convenience check
@@ -22,33 +23,52 @@ async function main() {
 
   console.log("Account balance:", (await deployer.getBalance()).toString());
 
-  const Token = await ethers.getContractFactory("Token");
-  const token = await Token.deploy();
-  await token.deployed();
+  // const Token = await ethers.getContractFactory("Token");
+  // const token = await Token.deploy();
 
-  console.log("Token address:", token.address);
+  const Nft_web3_explrer = await ethers.getContractFactory("NFT_WEB3_EXPLORER");
+  const nft_web3_explrer = await Nft_web3_explrer.deploy();
 
+  // await token.deployed();
+  await nft_web3_explrer.deployed();
+  // console.log("Token address:", token.address);
+  console.log("nft address",nft_web3_explrer.address)
   // We also save the contract's artifacts and address in the frontend directory
-  saveFrontendFiles(token);
+  // saveFrontendFiles(token,"Token");
+  saveFrontendFiles(nft_web3_explrer,"NFT_WEB3_EXPLORER");
 }
 
-function saveFrontendFiles(token) {
+function saveFrontendFiles(token,name) {
   const fs = require("fs");
   const contractsDir = path.join(__dirname, "..", "frontend", "src", "contracts");
 
   if (!fs.existsSync(contractsDir)) {
     fs.mkdirSync(contractsDir);
   }
+  const contract_address = path.join(__dirname, "..", "frontend", "src", "contracts","contract-address.json");
+  if(!fs.existsSync(contract_address)){
+    fs.writeFileSync(
+        path.join(contractsDir, "contract-address.json"),
+        JSON.stringify({ [name]: token.address }, undefined, 2)
+    );
+  }else{
+    let contentStr = fs.readFileSync(contract_address,{encoding:"utf8"});
+    let contentJson = JSON.parse(contentStr);
+    contentJson[name] = token.address;
+    fs.writeFileSync(contract_address,JSON.stringify(
+        contentJson,
+        undefined,
+        2),
+        {encoding:"utf8"});
+
+  }
+
+
+
+  const TokenArtifact = artifacts.readArtifactSync(name);
 
   fs.writeFileSync(
-    path.join(contractsDir, "contract-address.json"),
-    JSON.stringify({ Token: token.address }, undefined, 2)
-  );
-
-  const TokenArtifact = artifacts.readArtifactSync("Token");
-
-  fs.writeFileSync(
-    path.join(contractsDir, "Token.json"),
+    path.join(contractsDir, `${name}.json`),
     JSON.stringify(TokenArtifact, null, 2)
   );
 }
